requirements.txt
streamlit>=1.20
google-generativeai>=0.1.0      # pip package name may be google-generativeai
matplotlib
python-dotenv
requests
